Define trait, struct, and impl
A trait contains method signatures that implementing types must provide. If you want a field, then you need to define a setter and getter.

A struct is like a class or complex type, but without inheritance. Structs can define fields and methods and be implemented using the  syntax. We can implement structs with traits. This uses a composition paradigm instead of inheritance.